Restructured code
