+++ data/test.txt
@@ -1,0 +1,55 @@
+Country in depression
+Nation in despair
+One man seeking reasons everywhere
+Growing hate and anger
+The Führer's orders were precise
+Who was to be blamed and pay the price
+Wicked propaganda
+
+Turning neighbors into foes
+Soldiers of the third Reich searching homes
+And then the former friends are watching
+As they are rounded up one by one
+Times of prosecutions has begun
+
+Ever since it started
+On Kristallnacht '38
+When liberty died
+And truth was denied
+Sent away on train on a one way trip to hell
+Enter the gates, Auschwitz awaits
+
+When freedom burns
+The final solution
+Dreams fade away and all hope turns to dust
+When millions burn
+The curtain has fallen
+Lost to the world as they perish in flames
+
+There was a country in depression
+There was a nation in despair
+One man finding reasons everywhere
+Then there was raising hate and anger
+The Führer's orders still apply
+Who was to be blamed and sent to die
+
+Ever since it started
+On Kristallnacht '38
+When liberty died
+And truth was denied
+Sent away on train on a one way trip to hell
+Enter the gates, Auschwitz awaits
+
+When freedom burns
+The final solution
+Dreams fade away and all hope turns to dust
+When millions burn
+The curtain has fallen
+Lost to the world as they perish in flames
+
+When freedom burns
+The final solution
+Dreams fade away and all hope turns to dust
+When millions burn
+The curtain has fallen
+Lost to the world as they perish in flames
