+++ pyreport/errors.py
@@ -1,0 +1,2 @@
+class ReportError(Exception):
+    """Raised when report cannot be made."""
