+++ pyreport/utils.py
